requirements.txt
fastapi
streamlit
faiss-cpu
sentence-transformers
openai
pandas
numpy
uvicorn